fix: correct HXQ vec_dot offset formula in CUDA mmvq kernel
The offset term in vec_dot_hxq_affine_{g128,6}_q8_1 had wrong
parenthesization: `d8 * (scale * sumi + offset * s8)` applies an
extra d8 factor to the offset contribution.

Since s8 = d8 * sum(qs), the buggy formula computes
d8^2 * offset * sum(qs) instead of d8 * offset * sum(qs).

Fix: `d8 * scale * sumi + offset * s8`, matching the Q4_1/Q5_1
pattern where the min/offset term multiplies s8 directly without
an outer d8.

Symptom: correct output at ngl=0 (CPU), garbage at ngl>=1 (GPU)
when mixing CPU and GPU layers on T2000. Full-GPU (ngl=99 on 3090)
appeared correct because the systematic error was consistent across
all layers.

+++ ggml/src/ggml-cuda/vecdotq.cuh
@@ -748,7 +748,7 @@ static __device__ __forceinline__ float vec_dot_hxq_affine_g128_q8_1(
     const float d8 = __low2float(bq8->ds);
     const float s8 = __high2float(bq8->ds);  // sum of q8 values in this block
 
-    return d8 * (scale * sumi + offset * s8);
+    return d8 * scale * (float)sumi + offset * s8;
 }
 
 // HXQ affine 6-bit: 6-bit indices packed 4 per 3 bytes, scale+offset per group of 128
@@ -789,7 +789,7 @@ static __device__ __forceinline__ float vec_dot_hxq_affine_6_q8_1(
     const float d8 = __low2float(bq8->ds);
     const float s8 = __high2float(bq8->ds);
 
-    return d8 * (scale * sumi_idx + offset * s8);
+    return d8 * scale * (float)sumi_idx + offset * s8;
 }
 
 static __device__ __forceinline__ float vec_dot_q4_0_q8_1(
